refactor(projects): extract SelectComponent to separate file
- Create steps/select.tsx with Select, Identifiable, and SelectComponent
- Re-export from basic-information-form for backward compatibility
- Addresses code review feedback #4213
