+++ apps/web/core/components/features/projects/add-or-edit-project/steps/basic-information-form.tsx
@@ -5,7 +5,7 @@ import RichTextEditor from '../text-editor';
 import { cn } from '@/core/lib/helpers';
 import { CalendarIcon, X, Loader2 } from 'lucide-react';
 import { differenceInDays, format, isAfter, isEqual } from 'date-fns';
-import { CSSProperties, FormEvent, useCallback, useEffect, useMemo, useState, memo, useRef } from 'react';
+import { FormEvent, useCallback, useMemo, useState } from 'react';
 import { IStepElementProps } from '../container';
 import Image from 'next/image';
 import { isValidUrl } from '@/core/lib/utils';
@@ -13,14 +13,11 @@ import { ScrollArea, ScrollBar } from '@/core/components/common/scroll-area';
 import { useTranslations } from 'next-intl';
 import { useAuthenticateUser } from '@/core/hooks';
 import { useImageAssets } from '@/core/hooks/common/use-image-assets';
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem } from '@/core/components/common/command';
 import { Popover, PopoverContent, PopoverTrigger } from '@/core/components/common/popover';
-import { ChevronDown } from 'lucide-react';
 import { getInitialValue } from '@/core/lib/helpers/create-project';
 import { InputField } from '@/core/components/duplicated-components/_input';
 import { Button as ShadcnButton } from '@/core/components/duplicated-components/_button';
 import { Calendar } from '@/core/components/common/calendar';
-import { Checkbox } from '@/core/components/common/checkbox';
 
 interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
 	loading?: boolean;
@@ -535,303 +532,8 @@ export function DatePicker(props: IDatePickerProps) {
 	);
 }
 
-/**
- * Select (mono / multi)
- * With:
- * - Search
- * - Create new term if not found in the search results
- */
-
-export interface Identifiable {
-	id: string;
-	value: string | number;
-}
-
-type OnChange<IsMulti extends boolean> = IsMulti extends true ? (value: string[]) => void : (value: string) => void;
-type Selected<IsMulti extends boolean> = IsMulti extends true ? string[] : string | null;
-
-interface ISelectProps<IItem extends Identifiable, IsMulti extends boolean> {
-	options: IItem[];
-	selected: Selected<IsMulti>;
-	placeholder?: string;
-	selectTriggerClassName?: string;
-	selectTriggerStyles?: CSSProperties;
-	selectOptionsListClassName?: string;
-	multiple?: IsMulti;
-	onChange?: OnChange<IsMulti>;
-	renderItem?: (item: IItem, selected: boolean, active: boolean) => React.ReactNode;
-	renderValue?: (value: Selected<IsMulti>) => React.ReactNode;
-	searchEnabled?: boolean;
-	onCreate?: (newTerm: string) => void;
-	createLoading?: boolean;
-	showChevronDownIcon?: boolean;
-	alignOptionsList?: 'start' | 'center' | 'end';
-}
-
-/**
- * Performance-optimized Select component with comprehensive memoization
- *
- * - React.memo for component-level memoization with shallow comparison
- * - useCallback for stable event handlers preventing child re-renders
- * - useMemo for expensive calculations (filtering, height calculation)
- * - Optimized debouncing with ref-based timer management
- * - Memory-efficient string operations with minimal allocations
- * - Proper cleanup of timers and event listeners
- * - Stable reference handling for large datasets (285+ items)
- *
- * Performance characteristics:
- * - O(n) filtering complexity with early exit optimizations
- * - Debounced search prevents excessive filtering (150ms optimal)
- * - Memory footprint minimized through string caching
- * - Re-render frequency reduced by ~70% through strategic memoization
- *
- * @param props - Select component props with full backward compatibility
- * @returns Optimized Select component maintaining 100% API compatibility
- */
-function SelectComponent<T extends Identifiable, IsMulti extends boolean = false>(props: ISelectProps<T, IsMulti>) {
-	const {
-		options,
-		placeholder,
-		selectTriggerClassName,
-		selectTriggerStyles,
-		selected,
-		onChange,
-		renderItem,
-		multiple,
-		searchEnabled,
-		onCreate,
-		createLoading,
-		renderValue,
-		showChevronDownIcon = true,
-		selectOptionsListClassName,
-		alignOptionsList = 'start'
-	} = props;
-
-	// State management with performance considerations
-	const [searchTerm, setSearchTerm] = useState('');
-	const [debouncedSearchTerm, setDebouncedSearchTerm] = useState('');
-	const debounceTimerRef = useRef<NodeJS.Timeout | null>(null);
-
-	// Memoize boolean to prevent unnecessary recalculations
-	const isMulti = useMemo(() => multiple ?? false, [multiple]);
-
-	// Optimized debounce with proper cleanup and ref-based timer management
-	useEffect(() => {
-		// Clear existing timer
-		if (debounceTimerRef.current) {
-			clearTimeout(debounceTimerRef.current);
-		}
-
-		// Set new timer
-		debounceTimerRef.current = setTimeout(() => {
-			setDebouncedSearchTerm(searchTerm);
-		}, 150); // 150ms debounce - optimal balance between responsiveness and performance
-
-		// Cleanup function
-		return () => {
-			if (debounceTimerRef.current) {
-				clearTimeout(debounceTimerRef.current);
-			}
-		};
-	}, [searchTerm]);
-
-	// Cleanup timer on unmount
-	useEffect(() => {
-		return () => {
-			if (debounceTimerRef.current) {
-				clearTimeout(debounceTimerRef.current);
-			}
-		};
-	}, []);
-
-	/**
-	 * Memory-optimized search implementation with caching
-	 *
-	 * Performance optimizations:
-	 * 1. useMemo prevents unnecessary re-filtering on every render
-	 * 2. Early return for empty search terms
-	 * 3. Memory-efficient string operations with caching
-	 * 4. Optimized object allocations
-	 * 5. Stable options reference handling
-	 */
-	const filteredItems = useMemo(() => {
-		// Early return for empty search - no filtering needed
-		if (!debouncedSearchTerm?.trim()) {
-			return options || [];
-		}
-
-		const searchTermLower = debouncedSearchTerm.toLowerCase().trim();
-
-		// Performance optimization: pre-split search term once
-		const searchWords = searchTermLower.split(/\s+/).filter((word) => word.length > 0);
-
-		// Memory-efficient filtering with minimal allocations
-		return (options || []).filter((item) => {
-			// Type-safe string conversion with minimal allocations
-			const itemValue = String(item.value ?? '');
-			const itemId = String(item.id ?? '');
-
-			// Cache lowercased strings only when needed
-			const valueLower = itemValue.toLowerCase();
-			const idLower = itemId.toLowerCase();
-
-			// Enhanced search algorithm with optimized short-circuiting:
-			// 1. Direct substring match for simple searches (most common case)
-			// 2. Word-by-word matching for complex searches
-			// 3. Search across both ID and value fields
-
-			// Fast path: direct substring search (covers 90% of use cases)
-			if (valueLower.includes(searchTermLower) || idLower.includes(searchTermLower)) {
-				return true;
-			}
-
-			// Comprehensive path: word-by-word matching with early exit
-			return searchWords.every((word) => valueLower.includes(word) || idLower.includes(word));
-		});
-	}, [debouncedSearchTerm, options]);
-
-	// Memoized items calculation to prevent unnecessary recalculations
-	const items = useMemo(() => {
-		return searchEnabled && debouncedSearchTerm?.length ? filteredItems : (options ?? []);
-	}, [searchEnabled, debouncedSearchTerm?.length, filteredItems, options]);
-
-	// Memoized height calculation to prevent recalculation on every render
-	const listHeight = useMemo(() => {
-		const maxVisibleItems = 7;
-		const itemHeight = 2; // rem
-		return items?.length > maxVisibleItems ? '12rem' : `${items?.length * itemHeight}rem`;
-	}, [items?.length]);
-
-	// Memoized event handlers to prevent unnecessary re-renders of child components
-	const handleSearchTermChange = useCallback((value: string) => {
-		setSearchTerm(value);
-	}, []);
-
-	const handleItemSelect = useCallback(
-		(item: T) => {
-			if (isMulti) {
-				const newSelected = Array.isArray(selected) ? [...selected] : [];
-				const index = newSelected.indexOf(item.id);
-				if (index === -1) {
-					newSelected.push(item.id);
-				} else {
-					newSelected.splice(index, 1);
-				}
-				onChange?.(newSelected as any);
-			} else {
-				onChange?.(item.id as any);
-			}
-		},
-		[isMulti, selected, onChange]
-	);
-
-	const handleCreateNew = useCallback(() => {
-		onCreate?.(searchTerm);
-	}, [onCreate, searchTerm]);
-
-	return (
-		<div className="relative dark:bg-dark--theme-light">
-			<Popover>
-				<PopoverTrigger asChild>
-					<Button
-						style={selectTriggerStyles}
-						variant="outline"
-						role="combobox"
-						className={cn(
-							'flex justify-between items-center px-3 py-2 w-full h-10 text-sm text-left rounded-lg border dark:bg-dark--theme-light dark:border-white/20 dark:text-white',
-							selectTriggerClassName
-						)}
-					>
-						{renderValue ? (
-							renderValue(selected)
-						) : (
-							<span className={cn('capitalize', !selected?.length && 'text-gray-400 dark:text-gray-500')}>
-								{isMulti ? placeholder : options?.find((el) => el.id == selected)?.value || placeholder}
-							</span>
-						)}
-						{showChevronDownIcon ? (
-							<ChevronDown className="ml-2 w-4 h-4 opacity-50 shrink-0 dark:text-white" />
-						) : null}
-					</Button>
-				</PopoverTrigger>
-				<PopoverContent
-					className={cn(
-						'p-0 w-[var(--radix-popover-trigger-width)] dark:bg-dark--theme-light dark:border-white/20',
-						selectOptionsListClassName
-					)}
-					align={alignOptionsList}
-				>
-					<Command className="w-full dark:bg-dark--theme-light" shouldFilter={false}>
-						{searchEnabled && (
-							<CommandInput
-								placeholder={items?.length == 0 ? 'Type new ...' : 'Search ...'}
-								value={searchTerm}
-								onValueChange={handleSearchTermChange}
-								className="h-9 text-sm dark:bg-dark--theme-light dark:text-white dark:placeholder:text-gray-500"
-							/>
-						)}
-						<CommandEmpty>
-							{searchEnabled && items?.length == 0 && onCreate && (
-								<Button
-									type="button"
-									loading={createLoading}
-									onClick={handleCreateNew}
-									variant="outline"
-									className="w-full h-9 text-sm dark:border-white/20 dark:text-white hover:dark:bg-dark--theme"
-								>
-									Add new
-								</Button>
-							)}
-						</CommandEmpty>
-						<CommandGroup className="dark:bg-dark--theme-light">
-							<ScrollArea style={{ height: listHeight }}>
-								{items?.map((item, index) => (
-									<CommandItem
-										key={item?.id || `item-${index}`}
-										value={item?.id}
-										onSelect={() => handleItemSelect(item)}
-										className={cn(
-											'text-sm cursor-pointer rounded-sm px-2 py-1.5 dark:text-white dark:hover:bg-dark--theme',
-											isMulti && 'flex items-center gap-2'
-										)}
-									>
-										{renderItem ? (
-											renderItem(item, selected ? selected.includes(item.id) : false, false)
-										) : isMulti ? (
-											<>
-												<Checkbox
-													checked={selected?.includes(item.id)}
-													className="w-4 h-4 dark:border-white/20"
-												/>
-												<span className="capitalize dark:text-white">{item?.value ?? '-'}</span>
-											</>
-										) : (
-											<span className="capitalize dark:text-white">{item?.value ?? '-'}</span>
-										)}
-									</CommandItem>
-								))}
-								<ScrollBar className="-pl-7 dark:bg-dark--theme" />
-							</ScrollArea>
-						</CommandGroup>
-					</Command>
-				</PopoverContent>
-			</Popover>
-		</div>
-	);
-}
-
-/**
- * Performance-optimized Select component with React.memo
- *
- * Memoization strategy:
- * - Shallow comparison of props to prevent unnecessary re-renders
- * - Stable references maintained through useCallback and useMemo
- * - Memory-efficient filtering and caching
- *
- * @param props - Select component props
- * @returns Memoized Select component
- */
-export const Select = memo(SelectComponent) as typeof SelectComponent;
+export type { Identifiable } from './select';
+export { Select } from './select';
 
 /**
  * Show image or identifier letters
